Restructured editor and runtime features, changed namespaces.
