+++ Packages/dev.azmidi.oscmooth/Script/Editor/OSCmooth.Editor.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "OSCmooth.Editor",
+    "references": [
+        "OSCmooth"
+    ],
+    "includePlatforms": [
+        "Editor"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
