+++ Packages/dev.azmidi.oscmooth/Script/OSCmooth.asmdef
@@ -1,14 +1,3 @@
-{
-    "name": "OSCmooth",
-    "references": [],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
+﻿{
+	"name": "OSCmooth"
 }
