+++ Packages/dev.azmidi.oscmooth/Script/OSCmoothTypes.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 using static VRC.SDK3.Avatars.Components.VRCAvatarDescriptor;
 
-namespace Tools.OSCmooth.Types
+namespace OSCmooth.Types
 {
     [Serializable]
     public class OSCmoothLayers : ScriptableObject
@@ -22,6 +22,19 @@ namespace Tools.OSCmooth.Types
         {
             this.associate = associate;
         }
+    }
+
+    [Serializable]
+    public class OSCmoothParameter
+    {
+        // Default values
+        public float localSmoothness = 0.1f;
+        public float remoteSmoothness = 0.7f;
+        public string paramName = "NewParam";
+        public bool convertToProxy = true;
+        public int binarySizeSelection = 0;
+        public bool combinedParameter = false;
+        public bool isVisible = false;
     }
 }
